DPT-136 renamed lambda handler
